# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Security Triage API",
    "description": "Local development foundation. Scanner execution is not implemented yet.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/scans": {
      "post": {
        "tags": [
          "scans"
        ],
        "summary": "Create Scan",
        "description": "Queue a scan of a named npm package version or GitHub commit; the server builds the URL.",
        "operationId": "create_scan_api_v1_projects__project_id__scans_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanAccepted"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scans/{scan_id}": {
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "Get Scan",
        "description": "Status with timestamps, tool runs and the number of units per priority.",
        "operationId": "get_scan_api_v1_scans__scan_id__get",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanRead"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scans/{scan_id}/units": {
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "List Units",
        "description": "Location units in queue order (P1, P2, U, P3, P4, then unit_key).",
        "operationId": "list_units_api_v1_scans__scan_id__units_get",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scan Id"
            }
          },
          {
            "name": "tier",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "P1",
                    "P2",
                    "U",
                    "P3",
                    "P4"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tier"
            }
          },
          {
            "name": "tool",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "semgrep",
                    "codeql"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tool"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of items to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Maximum number of items to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitPage"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scans/{scan_id}/units/{unit_id}": {
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "Get Unit",
        "description": "Evidence record, reason, policy provenance and the raw findings of one unit.",
        "operationId": "get_unit_api_v1_scans__scan_id__units__unit_id__get",
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Unit Id"
            }
          },
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Body of every error response; matches FastAPI's own HTTPException shape."
      },
      "FindingRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tool": {
            "type": "string",
            "title": "Tool"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "raw_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Id"
          },
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "raw_rule_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Rule Id"
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          },
          "start_line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Line"
          },
          "end_line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Line"
          },
          "start_column": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Column"
          },
          "end_column": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Column"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "cwe_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cwe Ids"
          },
          "raw_result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Raw Result"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tool",
          "language",
          "raw_id",
          "rule_id",
          "raw_rule_id",
          "file_path",
          "start_line",
          "end_line",
          "start_column",
          "end_column",
          "message",
          "cwe_ids",
          "raw_result"
        ],
        "title": "FindingRead"
      },
      "GitHubSource": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "github",
            "title": "Kind"
          },
          "owner": {
            "type": "string",
            "maxLength": 39,
            "title": "Owner"
          },
          "repo": {
            "type": "string",
            "maxLength": 100,
            "title": "Repo"
          },
          "commit": {
            "type": "string",
            "maxLength": 40,
            "minLength": 40,
            "title": "Commit",
            "description": "Full commit SHA-1"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "owner",
          "repo",
          "commit"
        ],
        "title": "GitHubSource"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NpmSource": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "npm",
            "title": "Kind"
          },
          "package": {
            "type": "string",
            "maxLength": 214,
            "title": "Package",
            "examples": [
              "curling"
            ]
          },
          "version": {
            "type": "string",
            "maxLength": 256,
            "title": "Version",
            "description": "Exact SemVer version",
            "examples": [
              "0.2.0"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "package",
          "version"
        ],
        "title": "NpmSource"
      },
...
      "ScanAccepted": {
        "properties": {
          "scan_id": {
            "type": "string",
            "format": "uuid",
            "title": "Scan Id"
          },
          "snapshot_id": {
            "type": "string",
            "format": "uuid",
            "title": "Snapshot Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "scan_id",
          "snapshot_id",
          "status"
        ],
        "title": "ScanAccepted"
      },
      "ScanCreate": {
        "properties": {
          "source": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/NpmSource"
              },
              {
                "$ref": "#/components/schemas/GitHubSource"
              }
            ],
            "title": "Source",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "github": "#/components/schemas/GitHubSource",
                "npm": "#/components/schemas/NpmSource"
              }
            }
          },
          "profile": {
            "type": "string",
            "const": "command-injection-v0.1",
            "title": "Profile",
            "default": "command-injection-v0.1"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source"
        ],
        "title": "ScanCreate"
      },
      "ScanRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "reconciler_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reconciler Version"
          },
          "policy_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Version"
          },
          "snapshot": {
            "$ref": "#/components/schemas/SnapshotRead"
          },
          "tool_runs": {
            "items": {
              "$ref": "#/components/schemas/ToolRunRead"
            },
            "type": "array",
            "title": "Tool Runs"
          },
          "unit_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "P1",
                "P2",
                "U",
                "P3",
                "P4"
              ]
            },
            "type": "object",
            "title": "Unit Counts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "status",
          "profile",
          "created_at",
          "started_at",
          "finished_at",
          "error_message",
          "reconciler_version",
          "policy_version",
          "snapshot",
          "tool_runs",
          "unit_counts"
        ],
        "title": "ScanRead"
      },
      "SnapshotRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "source_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Kind"
          },
          "source_coordinate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Coordinate"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "provenance_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provenance Kind"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "source_kind",
          "source_coordinate",
          "sha256",
          "provenance_kind",
          "error_message"
        ],
        "title": "SnapshotRead"
      },
      "ToolRunRead": {
        "properties": {
          "tool": {
            "type": "string",
            "title": "Tool"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "tool_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Version"
          },
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "tool",
          "language",
          "status",
          "tool_version",
          "exit_code",
          "error_message"
        ],
        "title": "ToolRunRead"
      },
      "UnitDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "unit_key": {
            "type": "string",
            "title": "Unit Key",
            "description": "Canonical unit ID, comparable with the experiment CLI"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "start_line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Line"
          },
          "end_line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Line"
          },
          "start_column": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Column"
          },
          "end_column": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Column"
          },
          "sink_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sink Kind"
          },
          "argument_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Argument Role"
          },
          "mapping_status": {
            "type": "string",
            "title": "Mapping Status"
          },
          "priority": {
            "type": "string",
            "enum": [
              "P1",
              "P2",
              "U",
              "P3",
              "P4"
            ],
            "title": "Priority"
          },
          "decision_id": {
            "type": "string",
            "title": "Decision Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tools"
          },
          "matched_conditions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched Conditions"
          },
          "predicate_values": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Predicate Values"
          },
          "unknown_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unknown Fields"
          },
          "source_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Types"
          },
          "blocker_proof": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocker Proof"
          },
          "finding_evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Finding Evidence"
          },
          "policy_id": {
            "type": "string",
            "title": "Policy Id"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version"
          },
          "policy_sha256": {
            "type": "string",
            "title": "Policy Sha256"
          },
          "spec_sha256": {
            "type": "string",
            "title": "Spec Sha256"
          },
          "rule_claims_version": {
            "type": "string",
            "title": "Rule Claims Version"
          },
          "rule_claims_sha256": {
            "type": "string",
            "title": "Rule Claims Sha256"
          },
          "reconciler_version": {
            "type": "string",
            "title": "Reconciler Version"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingRead"
            },
            "type": "array",
            "title": "Findings"
          }
        },
        "type": "object",
        "required": [
          "id",
          "unit_key",
          "path",
          "start_line",
          "end_line",
          "start_column",
          "end_column",
          "sink_kind",
          "argument_role",
          "mapping_status",
          "priority",
          "decision_id",
          "reason",
          "tools",
          "matched_conditions",
          "predicate_values",
          "unknown_fields",
          "source_types",
          "blocker_proof",
          "finding_evidence",
          "policy_id",
          "policy_version",
          "policy_sha256",
          "spec_sha256",
          "rule_claims_version",
          "rule_claims_sha256",
          "reconciler_version",
          "findings"
        ],
        "title": "UnitDetail"
      },
      "UnitPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UnitSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "UnitPage",
        "description": "One page of a scan's location units in queue order."
      },
      "UnitSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "unit_key": {
            "type": "string",
            "title": "Unit Key",
            "description": "Canonical unit ID, comparable with the experiment CLI"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "start_line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Line"
          },
          "end_line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Line"
          },
          "start_column": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Column"
          },
          "end_column": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Column"
          },
          "sink_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sink Kind"
          },
          "argument_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Argument Role"
          },
          "mapping_status": {
            "type": "string",
            "title": "Mapping Status"
          },
          "priority": {
            "type": "string",
            "enum": [
              "P1",
              "P2",
              "U",
              "P3",
              "P4"
            ],
            "title": "Priority"
          },
          "decision_id": {
            "type": "string",
            "title": "Decision Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tools"
          }
        },
        "type": "object",
        "required": [
          "id",
          "unit_key",
          "path",
          "start_line",
          "end_line",
          "start_column",
          "end_column",
          "sink_kind",
          "argument_role",
          "mapping_status",
          "priority",
          "decision_id",
          "reason",
          "tools"
        ],
        "title": "UnitSummary"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}